[unit tests] [loaders] Refactor loaders (make hierarchy more consistent, remove duplication of code)
