+++ tests/ClassMetadataLoader/Fixture/Loader.php
@@ -1,6 +1,7 @@
 <?php
 namespace Kassko\DataMapperTest\ClassMetadataLoader\Fixture;
 
+use Kassko\DataMapper\ClassMetadata\ClassMetadata;
 use Kassko\DataMapper\ClassMetadataLoader;
 use Kassko\DataMapper\ClassMetadataLoader\LoadingCriteriaInterface;
 
@@ -9,5 +10,13 @@ class Loader extends ClassMetadataLoader\AbstractLoader
     function supports(LoadingCriteriaInterface $loadingCriteria)
     {
         throw new \RuntimeException('Method does not implemented.');
+    }
+
+    protected function doGetData(LoadingCriteriaInterface $loadingCriteria)
+    {
+    }
+
+    protected function doLoadClassMetadata(ClassMetadata $classMetadata, array $data)
+    {
     }
 }
